+++ Nos/Models/Author+CoreDataClass.swift
@@ -8,10 +8,13 @@
 
 import Foundation
 import CoreData
+import Dependencies
 import Logger
 
 @objc(Author)
 public class Author: NosManagedObject {
+    
+    @Dependency(\.currentUser) var currentUser
     
     var npubString: String? {
         publicKey?.npub
@@ -158,28 +161,6 @@ public class Author: NosManagedObject {
         return fetchRequest
     }
     
-    @MainActor @nonobjc class func inNetworkRequest(for author: Author? = nil) -> NSFetchRequest<Author> {
-        var author = author
-        if author == nil {
-            guard let currentUser = CurrentUser.shared.author else {
-                return emptyRequest()
-            }
-            author = currentUser
-        }
-        
-        let fetchRequest = NSFetchRequest<Author>(entityName: "Author")
-        fetchRequest.sortDescriptors = [NSSortDescriptor(keyPath: \Author.hexadecimalPublicKey, ascending: false)]
-        fetchRequest.predicate = NSPredicate(
-            format: "ANY followers.source IN %@.follows.destination " +
-                "OR hexadecimalPublicKey IN %@.follows.destination.hexadecimalPublicKey OR " +
-                "hexadecimalPublicKey = %@.hexadecimalPublicKey",
-            author!,
-            author!,
-            author!
-        )
-        return fetchRequest
-    }
-    
     /// Fetches all the authors who are further than 2 hops away on the social graph for the given `author`.
     static func outOfNetwork(for author: Author) -> NSFetchRequest<Author> {
         let fetchRequest = NSFetchRequest<Author>(entityName: "Author")
@@ -229,8 +210,8 @@ public class Author: NosManagedObject {
         print("Adding \(relay.address ?? "") to \(hexadecimalPublicKey ?? "")")
     }
     
-    func mute(context: NSManagedObjectContext) async throws {
-        guard let mutedAuthorKey = hexadecimalPublicKey, let currentAuthor = await CurrentUser.shared.author,
+    @MainActor func mute(viewContext context: NSManagedObjectContext) async throws {
+        guard let mutedAuthorKey = hexadecimalPublicKey, let currentAuthor = currentUser.author,
             mutedAuthorKey != currentAuthor.hexadecimalPublicKey else {
             return
         }
@@ -238,7 +219,7 @@ public class Author: NosManagedObject {
         print("Muting \(mutedAuthorKey)")
         muted = true
 
-        var mutedList = try await loadMuteList(context: context)
+        var mutedList = try await loadMuteList(viewContext: context)
 
         mutedList.append(mutedAuthorKey)
 
@@ -258,7 +239,7 @@ public class Author: NosManagedObject {
             try context.save()
         }
         // Publish the modified list
-        await CurrentUser.shared.publishMuteList(keys: Array(Set(mutedList)))
+        await currentUser.publishMuteList(keys: Array(Set(mutedList)))
     }
     
     func remove(relay: Relay) {
@@ -266,8 +247,8 @@ public class Author: NosManagedObject {
         print("Removed \(relay.address ?? "") from \(hexadecimalPublicKey ?? "")")
     }
 
-    func loadMuteList(context: NSManagedObjectContext) async throws -> [String] {
-        guard let currentAuthor = await CurrentUser.shared.author else {
+    @MainActor func loadMuteList(viewContext context: NSManagedObjectContext) async throws -> [String] {
+        guard let currentAuthor = currentUser.author else {
             throw CurrentUserError.authorNotFound
         }
         let request = currentAuthor.allPostsRequest(eventKind: .mute)
@@ -279,8 +260,8 @@ public class Author: NosManagedObject {
         }
     }
     
-    func unmute(context: NSManagedObjectContext) async throws {
-        guard let unmutedAuthorKey = hexadecimalPublicKey, let currentAuthor = await CurrentUser.shared.author,
+    @MainActor func unmute(viewContext context: NSManagedObjectContext) async throws {
+        guard let unmutedAuthorKey = hexadecimalPublicKey, let currentAuthor = currentUser.author,
             unmutedAuthorKey != currentAuthor.hexadecimalPublicKey else {
             return
         }
@@ -288,7 +269,7 @@ public class Author: NosManagedObject {
         print("Un-muting \(unmutedAuthorKey)")
         muted = false
 
-        var mutedList = try await loadMuteList(context: context)
+        var mutedList = try await loadMuteList(viewContext: context)
 
         mutedList.removeAll(where: { $0 == unmutedAuthorKey })
 
@@ -300,6 +281,6 @@ public class Author: NosManagedObject {
         }
 
         // Publish the modified list
-        await CurrentUser.shared.publishMuteList(keys: Array(Set(mutedList)))
+        await currentUser.publishMuteList(keys: Array(Set(mutedList)))
     }
 }