+++ Nos/Models/NoteParser.swift
@@ -17,20 +17,35 @@ enum NoteParser {
     static func parse(attributedText: AttributedString) -> (String, [[String]]) {
         cleanLinks(in: attributedText)
     }
-
-    /// Parses the content and tags stored in a note and returns an attributed text that can be used for displaying
-    /// the note in the UI.
+    
+    /// Parses the content and tags stored in a note and returns an attributed text with tagged entities replaced
+    /// with readable names.
     static func parse(content: String, tags: [[String]], context: NSManagedObjectContext) -> AttributedString {
         var result = replaceTaggedNostrEntities(in: content, tags: tags, context: context)
         result = replaceNostrEntities(in: result)
-        let linkedString = (try? result.findAndReplaceUnformattedLinks(in: result)) ?? result
         do {
             return try AttributedString(
-                markdown: linkedString,
+                markdown: result,
                 options: AttributedString.MarkdownParsingOptions(interpretedSyntax: .inlineOnlyPreservingWhitespace)
             )
         } catch {
             return AttributedString(stringLiteral: content)
+        }
+    }
+
+    /// Parses the content and tags stored in a note and returns an attributed text and list of URLs that can be used 
+    /// to display the note in the UI.
+    static func parse(content: String, tags: [[String]], context: NSManagedObjectContext) -> (AttributedString, [URL]) {
+        var result = replaceTaggedNostrEntities(in: content, tags: tags, context: context)
+        result = replaceNostrEntities(in: result)
+        let (cleanedString, urls) = result.extractURLs()
+        do {
+            return (try AttributedString(
+                markdown: cleanedString,
+                options: AttributedString.MarkdownParsingOptions(interpretedSyntax: .inlineOnlyPreservingWhitespace)
+            ), urls)
+        } catch {
+            return (AttributedString(stringLiteral: content), urls)
         }
     }