+++ Nos/Service/CurrentUser.swift
@@ -1,5 +1,5 @@
 //
-//  CurrentUser.shared.swift
+//  CurrentUser.swift
 //  Nos
 //
 //  Created by Christopher Jorgensen on 2/21/23.
@@ -24,9 +24,9 @@ enum CurrentUserError: Error {
 // swiftlint:disable type_body_length superfluous_disable_command
 class CurrentUser: NSObject, ObservableObject, NSFetchedResultsControllerDelegate {
     
-    @MainActor static let shared = CurrentUser(persistenceController: PersistenceController.shared)
-    
     @Dependency(\.analytics) private var analytics
+    @Dependency(\.persistenceController) private var persistenceController
+    @Dependency(\.pushNotificationService) private var pushNotificationService
     
     // TODO: it's time to cache this
     var keyPair: KeyPair? {
@@ -76,10 +76,10 @@ class CurrentUser: NSObject, ObservableObject, NSFetchedResultsControllerDelegat
     }
     
     // swiftlint:disable implicitly_unwrapped_optional
-    @MainActor var viewContext: NSManagedObjectContext
-    var backgroundContext: NSManagedObjectContext
+    @MainActor var viewContext: NSManagedObjectContext!
+    var backgroundContext: NSManagedObjectContext!
     
-    @Published var socialGraph: SocialGraphCache
+    @Published var socialGraph: SocialGraphCache!
     
     var relayService: RelayService! {
         didSet {
@@ -104,11 +104,11 @@ class CurrentUser: NSObject, ObservableObject, NSFetchedResultsControllerDelegat
     
     private var authorWatcher: NSFetchedResultsController<Author>?
 
-    @MainActor init(persistenceController: PersistenceController) {
+    @MainActor override init() {
+        super.init()
         self.viewContext = persistenceController.viewContext
         self.backgroundContext = persistenceController.newBackgroundContext()
         self.socialGraph = SocialGraphCache(userKey: nil, context: backgroundContext)
-        super.init()
         if let privateKeyData = KeyChain.load(key: KeyChain.keychainPrivateKey) {
             Log.info("CurrentUser loaded a private key from keychain")
             let hexString = String(decoding: privateKeyData, as: UTF8.self)
@@ -156,7 +156,7 @@ class CurrentUser: NSObject, ObservableObject, NSFetchedResultsControllerDelegat
             
             Task(priority: .background) {
                 let dbStatistics = try await backgroundContext.perform {
-                    try PersistenceController.databaseStatistics(from: self.backgroundContext)
+                    try self.persistenceController.databaseStatistics(from: self.backgroundContext)
                 }
                 analytics.databaseStatistics(dbStatistics)
             }
@@ -187,9 +187,12 @@ class CurrentUser: NSObject, ObservableObject, NSFetchedResultsControllerDelegat
     }
 
     @MainActor func subscribe() async {
+        guard let key = publicKeyHex, let author else {
+            return
+        }
         
         let overrideRelays: [URL]?
-        let userRelays = author?.relays ?? Set()
+        let userRelays = author.relays 
         if userRelays.isEmpty {
             overrideRelays = Relay.allKnown
                 .compactMap {
@@ -201,28 +204,28 @@ class CurrentUser: NSObject, ObservableObject, NSFetchedResultsControllerDelegat
             overrideRelays = nil
         }
         
-        // Always listen to my changes
-        if let key = publicKeyHex, let author {
-            // Close out stale requests
-            if !subscriptions.isEmpty {
-                await relayService.decrementSubscriptionCount(for: subscriptions)
-                subscriptions.removeAll()
-            }
-            
-            // Subscribe to our own events of all kinds.
-            let latestRecievedEvent = try? viewContext.fetch(Event.lastReceived(for: author)).first
-            let allEventsFilter = Filter(authorKeys: [key], since: latestRecievedEvent?.receivedAt)
-            subscriptions.append(await relayService.openSubscription(with: allEventsFilter))
-            
-            // Always make a one time request for the latest contact list
-            let contactFilter = Filter(
-                authorKeys: [key],
-                kinds: [.contactList],
-                limit: 1,
-                since: author.lastUpdatedContactList
-            )
-            subscriptions.append(await relayService.openSubscription(with: contactFilter, to: overrideRelays))
+        // Close out stale requests
+        if !subscriptions.isEmpty {
+            await relayService.decrementSubscriptionCount(for: subscriptions)
+            subscriptions.removeAll()
         }
+        
+        // Subscribe to our own events of all kinds.
+        let latestRecievedEvent = try? viewContext.fetch(Event.lastReceived(for: author)).first
+        let allEventsFilter = Filter(authorKeys: [key], since: latestRecievedEvent?.receivedAt)
+        subscriptions.append(await relayService.openSubscription(with: allEventsFilter))
+        
+        // Always make a one time request for the latest contact list
+        let contactFilter = Filter(
+            authorKeys: [key],
+            kinds: [.contactList],
+            limit: 1,
+            since: author.lastUpdatedContactList
+        )
+        subscriptions.append(await relayService.openSubscription(with: contactFilter, to: overrideRelays))
+        
+        // Listen for notifications
+        await pushNotificationService.listen(for: self)
     }
     
     private var friendMetadataTask: Task<Void, any Error>?