+++ Nos/Views/ReportMenu.swift
@@ -81,7 +81,7 @@ struct ReportMenuModifier: ViewModifier {
     func mute(author: Author) {
         Task {
             do {
-                try await author.mute(context: viewContext)
+                try await author.mute(viewContext: viewContext)
             } catch {
                 Log.error(error.localizedDescription)
             }
@@ -144,13 +144,14 @@ extension View {
 
 struct ReportMenu_Previews: PreviewProvider {
     
+    static var previewData = PreviewData()
     static var previews: some View {
         StatefulPreviewContainer(false) { binding in
             VStack {
                 Button("Report this") { 
                     binding.wrappedValue.toggle()
                 }
-                .reportMenu(binding, reportedObject: .note(PreviewData.imageNote))
+                .reportMenu(binding, reportedObject: .note(previewData.imageNote))
             }
         }
     }