+++ Nos/Views/ThreadView.swift
@@ -62,22 +62,16 @@ struct ThreadView: View {
 }
 
 struct ThreadView_Previews: PreviewProvider {
+    static var previewData = PreviewData()
     static var persistenceController = PersistenceController.preview
     static var previewContext = persistenceController.container.viewContext
-    static var relayService = RelayService(persistenceController: persistenceController)
+    static var relayService = previewData.relayService
     static var router = Router()
     
     static var emptyPersistenceController = PersistenceController.empty
     static var emptyPreviewContext = emptyPersistenceController.container.viewContext
-    static var emptyRelayService = RelayService(persistenceController: emptyPersistenceController)
-
-    static var currentUser: CurrentUser = {
-        let currentUser = CurrentUser(persistenceController: persistenceController)
-        currentUser.viewContext = previewContext
-        currentUser.relayService = relayService
-        Task { await currentUser.setKeyPair(KeyFixture.keyPair) }
-        return currentUser
-    }()
+    static var emptyRelayService = previewData.relayService
+    static var currentUser = previewData.currentUser
     
     static var alice: Author = {
         let author = Author(context: previewContext)